Grouped types and cleaned up
