+++ src/index.ts
@@ -1,2 +1,1 @@
-export * as types from './types'
 export { zbd } from './zbd';
