+++ src/types/index.ts
@@ -1,0 +1,13 @@
+export * from './charges';
+export * from './gamertag';
+export * from './index';
+export * from './internalTransfers';
+export * from './keysend';
+export * from './lightning';
+export * from './misc';
+export * from './payments';
+export * from './staticCharges';
+export * from './wallet';
+export * from './withdrawal';
+export * from './zbdType';
+
