+++ src/types/wallet.ts
@@ -1,0 +1,9 @@
+
+export interface WalletDataResponseType {
+    data: {
+      balance: string;
+      unit: string;
+    }
+    message: string;
+    success: boolean;
+  }
