+++ src/types/zbdType.ts
@@ -1,0 +1,5 @@
+export interface ZBD {
+    apikey: string;
+    apiBaseUrl: string;
+    axiosInstance: any;
+}
